pyproject.toml
[project]
name = "recipe-book"
version = "0.1.0"
description = "An app to store your recipes."
readme = "README.md"
authors = [
    { name = "Kishan Patel", email = "kishanpatel789@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"
markers = [
    "focus: Run focused tests"
]

[tool.coverage.report]
omit = [
    "tests/*",
    "*/migrations/*",
    "*/management/*",
]
